refactor: modules/wallet/utils
