+++ src/modules/wallet/utils/getAddEthereumChainParameters.ts
@@ -1,0 +1,55 @@
+import { ChainId, getChainName } from '@dcl/schemas/dist/dapps/chain-id';
+import { getChainConfiguration } from '../../../lib/chainConfiguration';
+import { AddEthereumChainParameters } from '../types';
+
+
+
+export function getAddEthereumChainParameters(
+  chainId: ChainId
+): AddEthereumChainParameters {
+  const hexChainId = '0x' + chainId.toString(16);
+  const chainName = getChainName(chainId)!;
+  const config = getChainConfiguration(chainId);
+  switch (chainId) {
+    case ChainId.MATIC_MAINNET:
+      return {
+        chainId: hexChainId,
+        chainName,
+        nativeCurrency: {
+          name: 'MATIC',
+          symbol: 'MATIC',
+          decimals: 18
+        },
+        rpcUrls: ['https://rpc-mainnet.maticvigil.com/'],
+        blockExplorerUrls: ['https://polygonscan.com/']
+      };
+    case ChainId.MATIC_MUMBAI:
+      return {
+        chainId: hexChainId,
+        chainName,
+        nativeCurrency: {
+          name: 'MATIC',
+          symbol: 'MATIC',
+          decimals: 18
+        },
+        rpcUrls: ['https://rpc-mumbai.maticvigil.com/'],
+        blockExplorerUrls: ['https://mumbai.polygonscan.com/']
+      };
+    case ChainId.ETHEREUM_MAINNET:
+    case ChainId.ETHEREUM_ROPSTEN:
+    case ChainId.ETHEREUM_RINKEBY:
+    case ChainId.ETHEREUM_KOVAN:
+    case ChainId.ETHEREUM_GOERLI:
+      return {
+        chainId: hexChainId,
+        chainName,
+        nativeCurrency: {
+          name: 'Ether',
+          symbol: 'ETH',
+          decimals: 18
+        },
+        rpcUrls: [config.rpcURL],
+        blockExplorerUrls: ['https://etherscan.io']
+      };
+  }
+}
