+++ src/modules/wallet/utils/transactionEvents.ts
@@ -1,0 +1,2 @@
+import { EventEmitter } from 'events'
+export const transactionEvents = new EventEmitter()
