+++ src/modules/wallet/utils/types.ts
@@ -1,0 +1,12 @@
+export enum TransactionEventType {
+  ERROR = 'error',
+  SUCCESS = 'success'
+}
+
+export type TransactionEventData<T extends TransactionEventType> = {
+  type: T
+} & (T extends TransactionEventType.ERROR
+  ? { error: Error }
+  : T extends TransactionEventType.SUCCESS
+  ? { txHash: string }
+  : {})
